+++ modules/SimpleMap/SimpleMapWindow.hh
@@ -3,17 +3,18 @@
 #define SIMPLE_MAP_WINDOW_H_
 
 #include "DataManager.hh"
-#include "MapEditor.hh"
+#include "Lib.hh"
+#include "MapEditorBase.hh"
 #include "SimpleMap.hh"
 #include "SimpleMapConfig.hh"
 #include "WindowBase.hh"
 
 class SimpleMapWindow : public WindowBase {
   public:
-    float      bsize;
-    float      padding;
-    SimpleMap* simple_map;
-    MapEditor* map_editor;
+    float          bsize;
+    float          padding;
+    SimpleMap*     simple_map;
+    MapEditorBase* map_editor;
 
     SimpleMapWindow();